rename `datafrog` module to `kind_inference`
