+++ components/lark-type-check/src/full_inference/query_definition.rs
@@ -1,4 +1,4 @@
-use crate::full_inference::datafrog;
+use crate::full_inference::kind_inference;
 use crate::full_inference::resolve_to_full_inferred::ResolveToFullInferred;
 use crate::full_inference::type_checker::FullInferenceStorage;
 use crate::full_inference::FullInference;
@@ -39,7 +39,7 @@ crate fn full_type_check(
 
     type_checker.check_fn_body();
 
-    let perm_kinds = datafrog::inference(&type_checker, &type_checker.storage.constraints);
+    let perm_kinds = kind_inference::inference(&type_checker, &type_checker.storage.constraints);
 
     let mut unresolved_variables = vec![];
     let inferred_results = type_checker
